reorganize: settings
